+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `--disable-pager` disable pagers.
 - Add `--script-mode` to retain stdout/stderr as separate
   streams but without raw mode on the terminal.
+- Added `--disable-raw` to explicitly disable raw mode.
 - Improved `EINTR` handling.
 - Improve compatibility with older rust versions.